moved out TestScheduler extensions in its own file
