+++ Firebase/ios/BonfireTests/TestScheduler+Extensions.swift
@@ -1,0 +1,18 @@
+import Foundation
+import RxSwift
+import RxTests
+
+extension TestScheduler {
+    func singleEventAndComplete<Element>(element: Element) -> Observable<Element> {
+        return self.createColdObservable([
+            next(0, element),
+            completed(0)
+            ]).asObservable()
+    }
+
+    func singleEventAndHang<Element>(element: Element) -> Observable<Element> {
+        return self.createHotObservable([
+            next(0, element)
+            ]).asObservable()
+    }
+}
